modularizing the implementation while including only test cases.
