+++ java-copy-concepts/src/main/java/com/baeldung/copy/model/Item.java
@@ -1,0 +1,35 @@
+package com.baeldung.copy.model;
+public class Item implements Cloneable{
+    private String name;
+    private double price;
+
+    public Item(Item that) {
+        this(that.name, that.price);
+    }
+
+    public Item(String name, double price) {
+        this.name = name;
+        this.price = price;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public double getPrice() {
+        return price;
+    }
+
+    public void setPrice(double price) {
+        this.price = price;
+    }
+
+    @Override
+    public Object clone() throws CloneNotSupportedException {
+        return super.clone();
+    }
+}
